Simplify AST
In this commit, we add an intermediate stage to the compilation
pipeline, in between parsing and node registration.
This stage is the "AST simplification stage".
In this stage, we transform Identifier and Dot expressions into a more
general NameExpression node.
This eliminates the need to deal with Identifiers and Dots separately
(which is a nuisance, since the logic is usually more or less
identical).

This commit is very squashed. To see the original individual subcommits,
please check out the `branch-archive/transform-ast-to-sst` tag.