Forbid variant access until type declaration is complete
For example...

```pamlihu
type Nat {
    .O: Nat,
    .S(obvious: Eq(Nat, Nat.O, Nat.O), pred: Nat): Nat,
}
```

...is illegal, since `Nat.O` is referenced before the declaration
of `Nat` is complete (i.e., in type of the `obvious` parameter).

Note that only _variant_ access is forbidden.
Access to the _type itself_ within variant declarations is permitted.
For example,

```pamlihu
type List(T: Type) {
    .Nil(T: Type): List(T),
    .Cons(T: Type, car: T, cdr: List(T)): List(T),
}
```

Since the `cdr` param references the `List` _type_, and not one of its
variants, this is legal.

This restriction is to help guarantee soundness.
In the future, if we can prove that programs will still be sound
without this restriction, we will remove it.