Find bug 1 cause
Bug 1 is that the `Nat` in `type Nat {..., .S(n: >>Nat<<)}` (i.e.,
the one surrounded in `>> <<`) is being reported as an illegal type.

The first cause of this is that the `remove_id` module incorrectly
removes the id from nodes (i.e., it removes data that should not be
removed).
Specifically, one instance of this error is when the `remove_id` module
removes the `db_index` from the `NameExpression` node.
As a result, when `try_shift_with_cutoff` tries to register the shifted
node with the registry, the registry considers the shifted node and the
original node to be equivalent (since `remove_id::NameExpression` does
not have a `db_index`).
Thus, it returns the original node (since it doesn't register a new
node).

There are probably other causes, but until this one is fixed, it is
inefficient to determine them.
So, we will address this one first.