Fix bug in normalize_params_and_leave_params_in_context
Previously, we calculated the type of the normalized parameter by
taking calling `type_check_param` (which adds the parameter to the top
of the context) and then retrieving the type from the context.
However, when retrieving the type, the retrieved type will be upshifted
by one level, to account for the fact that the context now includes
the param.
However, we need to define the param's type _before_ it's added to the
context, so we need do not want the upshift.
This commit fixes this by downshifting the retrieved type.