Fix Fun return type bug
Previously, we forgot to upshift a Fun's return type before comparing
it against its body's type.
This resulted in an off-by-one error, since the original return type
was defined in a context _without_ the Fun entry (since the return type
is forbidden to recursively refer to the Fun), and the body was defined
in a context _with_ the Fun entry (since recursion within the body
_is_ allowed).
This commit fixes this.