Replace WasNoOp with WasSyntacticNoOp
Previously, WasNoOp denoted whether a substitution was a _semantic_
no-op.
Now it denotes whether a substitution was a _syntactic_ no-op.
The reason we care about syntactic no-ops is that we want to continue
substituting if the last substitution was syntactically significant
(i.e., it was not a syntactic no-op), even if it was semantically
insignificant.
This should produce better error messages.