Reduce problem
The major issue is that we are not fusing matchees with their
destructured expressions.

For example, if we have

```pamlihu
match a {
    .O => X(...),
    .S(a') => Y(...),
}
```

...then we should fuse `a` with `Nat.O` in `X`,
and `a` with `Nat.S(a')` in `Y`, respectively.

Of course, we need to figure out how to do this.
It won't be trivial like the matchee type fusion, since this could
result in forward references, so we need to figure out how to work
around that.

By "forward references", I mean, for example

```pamlihu
fun f(a: Nat, H: P(a)): Q(a) {
    match a {
        .O => ...,
        .S(a') => PROBLEM,
    }
}
```

In `PROBLEM`, we need to fuse `a` with `Nat.S(a')`, which means
`H` should have a type `P(Nat.S(a'))`. However, `a'` is not in scope
when `H` is declared, so we need to figure out how to work around that.