Rename parameterized matchee type to use case param names
Previously, it used the names of the forall params. Since we define
semantics in terms of De Bruijn indices (not names), this didn't affect
evaluation or type checking.
However, it led to unexpected error messages, since the user would
expect the parameterized matchee type to be, well,
correctly pameterized 😂
This commit changes the names to match that of the case params, so
the error messages are more intuitive.