Find one source of bug
The problem is_left_inclusive_subterm_of_right doesn't have a cutoff.
For example, consider

Expression 1 := `x<0>` (De Bruijn indices are in angle brackets).
Expression 2 := `forall(y: Z<47>) { y<0> }`.

Then is we ask "Is Expression 1 a subterm of Expression 2?", the
current function returns true, but it should return false.
The problem is that the function doesn't have a cutoff, so it mistakes
the bound variable `y<0>` for the free variable `x<0>`.