Fix is_left_inclusive_subterm_of_right
Previously, the function erroneously neglected to upshift left when
checking if left was a subterm of one of right's subterms.
However, in cases where the right subterm in question had a different
context length, this caused the function to return incorrect results.
Please see the previous commit message for more info about this bug.