Fix get_type_of_match_case bug
The issue was that `evaluate_well_typed_expression` was being called on
`case_output_id`, but before we checked in `case_output_id` was actually
well-typed. In cases where `case_output_id` turned out to be ill-typed,
this could potentially lead to a panic, like it previously did in the
failing test case.
Now, we call `evaluate_possibly_ill_typed_expression` instead, which
fixes the issue.