Fix underscore error
Previously, we skipped pushing to the context when we encountered
underscore identifiers.
This had the desired effect of preventing the user from referencing
identifier with underscore names.
However, it also erroneously offset the context.

Now, we push to the context, but we add a permantent restriction to
the pushed entry.