Add failing fun rec test
This test tests the fun recursion validator's context cleanup.
It intentionally tries to pollute the context by causing an early
exit (i.e., `return Err(...)`) by featuring an possibly invalid
expression that turns out to have an illegal fun rec error.
If the context cleanup is incorrectly implemented, this will cause the
cleanup to be skipped (since it returns early).

Since the test failed, the context cleanup implementation is incorrect.