Rename "unbindable" to "symbolically invalid"
This is to be consistent to "untypecheckable", which implies a failure
at some phase in the pipeline _before_ the typechecker. That is, with
an untypecheckable expression, we can't even _start_ type checking it.

However, an unbindable expression is something we construct _after_ we
try (and fail) to bind it. Thus, this leads to 2 uses of "un-":

- "untypecheckable" - In this case, "un-<X>" means we can't even start
  "<X>" (due to a failure upstream in the pipeline).
- "unbindable" - In this case, "un-<X>" means we attempted "<X>" and it
  failed.

This is a bit confusing, so in this commit, we choose the former
interpretation ("untypecheckable").
As a result, "unbindable" is no longer a compatible name, so we renamed
it to "symbolically invalid" to reflect the fact that we _already_
attempted the binding and concluded that it was invalid.