Lazily substitute in types right before comparing
Previously, we eagerly substituted terms when type checking match cases.
However, that leads to many complications, such as difficulty correctly
substituting in the match case output without evaluating the output and
overlooking type errors in the process.

Now, we only perform substitutions right before we compare two types
to check for assignability.

This commit squashes many commits from the deleted
use-lazy-substitution-when-checking-match-case branch.
To see the individual commits, please checkout the
branch-archive/use-lazy-substitution-when-checking-match-case tag.