TODO Squash: Simplify and fix the evaluator
Previously, it was popping `n` params from the context even if not all
`n` params were added (where `n` is the arity of a Fun).
I'm still not certain why this resulted in a bug, since in theory,
the "not all `n` params added" case should never happen, since
`eval_possibly_ill_typed_expression` was actually only called,
in practice, with well-typed expressions, as far as I know.

But I'm sleepy so I'm not going to worry about it too much right now.

The important thing is the tests pass once again! Finally! uicai