Determine current_file_id without using spans
In theory, this tight coupling shouldn't be a problem--an AST node's
span should always have the file id of the file it was written in.
However, in my opinion, it violates the separation of concerns between
error message information (which is what spans are for) and the actual
binding.
This commit explicitly tracks the current file id, so we don't need to
infer it from spans.